Renaming the arriving sensor.
